# Increase the problem size for string benchmarks