… benchmarks/str-gather.py