Bump the problem size for string benchmarks to 10**8, which matches the
numeric benchmarks. Previously, we used smaller problem sizes for string
benchmarks because they were slower. Regular argsort was optimized in
#671 and coargsort wasn't actually that slow so we can bump sort problem
size for them. str-gather itself is fast, but the cast was slow (see
#675) so switch to using random data to avoid that issue. The old goupby
problem size wasn't bad, but bump it to be consistent with other codes.
This also makes the string size consistent across benchmarks by
increasing the max length to 16 for argsort/coargsort.

Here's the rates with the old and new problem sizes on 16-node-cs-hdr:

| benchmark       | before      | after       |
| --------------- | ----------: | ----------: |
| 2-arr goupby    |  0.12 GiB/s |  0.64 GiB/s |
| gather          |  5.42 GiB/s | 12.54 GiB/s |
| argsort         |  0.01 GiB/s |  0.86 GiB/s |
| 2-arr coargsort |  0.01 GiB/s |  0.51 GiB/s |

For some of these benchmarks (especially sorting) there is a non-trivial
fixed overhead regardless of problem size. So the actual time to run
some of these hasn't changed drastically because most of the time was
spent in that initial set up, not doing the core operation. Here's the
times to demonstrate that:

| benchmark       | before      | after       |
| --------------- | ----------: | ----------: |
| 2-arr goupby    | 20.96 sec   | 39.13 sec   |
| gather          |  1.37 sec   |  5.94 sec   |
| argsort         |  6.28 sec   | 15.52 sec   |
| 2-arr coargsort | 13.18 sec   | 26.14 sec   |

Part of #659