@ronawho 2021-02-16
@reuster986 I'm also wondering if we should unify the string lengths across these benchmarks. argsort/coargsort use `random_strings_uniform(1, 8..)` and gather/groupby use `random_strings_uniform(1, 16...)`