@reuster986 2021-02-17
This will be great to bring the string benchmarks in line with the numeric ones. 

Regarding unifying string lengths, I'm thinking we should go with `random_strings_uniform(1, 16, ...)` across the board, so that average string length will be 8 and will be about the size of an equivalent numeric array (or two numeric arrays, if you consider the offsets).