@reuster986 2021-02-17
@ronawho I just merged @hokiegeek2 's fix for mypy, so a rebase should do the trick.