@ronawho 2021-02-17
Rebased and updated argsort/coargsort to use `random_strings_uniform(1, 16, ...)` 

Rates/times in PR description updated to reflect the new length (the rate is slightly higher)